Start to implement voter but have to work on authentication first

+++ src/AppBundle/Controller/VimeoController.php
@@ -98,6 +98,8 @@ class VimeoController extends Controller
         $em = $this->getDoctrine()->getManager();
         $vid = $em->getRepository(Vimeo::class)->find($id);
 
+        $this->denyAccessUnlessGranted('delete', $vid, 'Ce média est utilisé par une galerie, suppression impossible');
+
         $em->remove($vid);
         $em->flush();
